requirement.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cors==1.0.1
Django==5.1.1
djangorestframework==3.15.2
filelock==3.16.1
future==1.0.0
gevent==24.10.3
greenlet==3.1.1
idna==3.10
mysql-connector-python==9.1.0
mysqlclient==2.2.6
pillow==11.0.0
pycparser==2.22
PySocks==1.7.1
requests==2.32.3
requests-file==2.1.0
setuptools==75.2.0
sqlparse==0.5.1
tldextract==5.1.2
tzdata==2024.2
urllib3==2.2.3
zope.event==5.0
zope.interface==7.1.1
